Added Dog Model
Introduced Dog model which includes the name, temperament, image url and a type representing the life span.

LifeSpanYear is the type used for the life span data. When comparing LifeSpanYear, the average is used. If the life span is a constant number then it is used to compare with the average between the minimum and maximum of a life span that is in a range.