Added UIImageView fetch image extension and dequeTableCell UITableView extension
Added a UIImageView fetch image extension so abstract the use of the 3rd party library Kingfisher. This would make it easier to change to a different 3rd party library or use a bespoke implementation for image fetching in the future since you only need to change the calling code via the UIImageView extension.

Added a method for dequeuing a table view cell using the reuseIdentifier based on the return type expected.