Added DogListService, Models, Errors and processing logic
DogListService is the abstraction layer that the ViewModel will use to be able to get the data from the API. This way the ViewModel won't have any logic of the networking stack, but will only focus on getting the future of either an array of Dog or an error.

DogListResponse is the Decodable model used for the API response. This is different from the Dog model which the app is used. The goal is to be able to modify and iterate over the API shape and it wouldn't have to change all throughout usages in the app. Changes only need to apply in the conversion which is in Dog+Response extension