Added sort behavior
Sorting is done based on the average life span years provided. Added a UI bar button item that shows an action sheet for a user to pick sorting the data by ascending / descending order. Sorting is done on the existing items. It is done on a background thread for cases where sorting takes longer so as not to block the UI.